pipeline/news-generator/requirements.txt
openai>=1.40.0
feedparser>=6.0.0
pyyaml>=6.0
pillow>=10.0.0
